This PS adds a promenade yaml that templates the IPs and promenade image.
The promenade yaml will be used by Jenkins and the IPs will be dynamic
so we cannot have them hard-coded.  The version of the Promenade image is also
templated to allow for the Jenkins major.minor.patch tag to replace it.
Creating a generated_configs directory that will be used to store the generated
 yaml and script files.