Use restartPolicy: Never for kube-dns
Sometimes Calico appears ready before it is able to provide proper
networking for pods.  When this happens to the kube-dns pod, its
process will repeatedly crash and be restarted, but with
restartPolicy: Always, its networking will not be refreshed.

restartPolicy: Never causes the pod to be deleted, which in turn causes
the Kubernetes controller manager to create a new pod, which will
receive fresh networking.

Change-Id: Id0aa00c50b19a4c5902f6d27855858c992196bda