Tolerate unready endpoints for apiserver service
This avoids a race condition that can occur while joining nodes when
another node is down.

The service is really a discovery service that is primarily used by the
haproxy chart to find apiserver endpoints, so it's reasonable for it to
include all endpoints (HAProxy will monitor health on its own).

Change-Id: I09dd66883af4ff91f9d13432137094fdf876ca83