[gate] Mask AVX2 to prevent 16.04 on 16.04 VMs boot crashes
16.04 on 16.04 VMs will crash (illegal instruction in
raid6_avx21_gen_syndrome) on boot if the host has AVX2 and we pass
that through.

The issue seems to be the guest kernel sees presence of AVX2 so raid6
module makes use of it - though it's not enabled (different bits).

Until this is resolved in the host hypervisor (kvm) and/or guest
kernel, mask out AVX2 to prevent crashes in L1/L2 VMs.

Change-Id: I0ca8edb8f62f9f2e96aa5e265bac631c346d0eac