Update Deckhand version
A recent Promenade change [0] introduced user-context tracing support in
Promenade, which requires the passing of an additional parameter to the
Deckhand API client. The Deckhand client was updated to handle the
additional parameter [1], but the Deckhand version Promenade uses was
not updated. This change uplifts the version of Deckhand in
requirements.txt for Promenade to a version that supports the
aforementioned, additional, user-context parameter.

[0] https://review.openstack.org/634071
[1] https://review.openstack.org/634068

Change-Id: Iaea328b32f95262e9e81be957893b7a50ac295fa