CoreDNS: remove external liveness dependency
This removes an external dependency from the default
liveness check.  Previously the coredns liveness probe
would attempt to resolve both an internal and external DNS
name to ensure DNS is working properly.  In practice, however,
external DNS resolution errors tend to be the result of
external networking issues, which won't be solved by deleting
and recreating the coredns service pod (which is the result
of the failed liveness probe).

If extneral name resolution is desired, it can still be
specified in the coredns chart override list.

Change-Id: Ife0d8ebab7bd298e1429bc23140c291e129ac9f5