Change CoreDNS from Deployment to DaemonSet
CoreDNS is critical to cluster operations, and is also a very lightweight
service.  This change makes the CoreDNS service deployable as a Deployment
(current behavior), a DaemonSet, or both simultaneously.  This allows
DNS to be easily configured to run on all nodes (or all control
plane nodes if desired) for high availability of the service and
resiliency of the cluster.

The "deplyoment and daemonset" behavior can be used to provide an
uninterrupted migration path from a Deployment-based environment
to a DaemonSet-based one.

Change-Id: I58c3e62ac4892a4d9374d99eefe8055865cebf1e