[etcd] improve etcd resiliency
This PS adds config param to reject reconfiguration request
that could cause quorum loss. This param is already present
in the etcd anchor pods. More details here [0]

Strict reconfig check mode:

The best practice of adding new members is to configure a single member
at a time and verify it starts correctly before adding more new members.
This step by step approach is very important because if newly added
members is not configured correctly (for example the peer URLs are incorrect),
the cluster can lose quorum. The quorum loss happens since the newly
added member are counted in the quorum even if that member is not 
reachable from other existing members. Also quorum loss might happen if
there is a connectivity issue or there are operational issues.

[0] https://github.com/etcd-io/etcd/blob/master/Documentation/op-guide/configuration.md#--strict-reconfig-check

Change-Id: I2261842d21cd051f4790d495b45ff7650f99b5e8