Armada metric output for genesis
This change has passed the Promenade resiliency gate.

Armada recently added support for generating metric output [0] [1].
This adds support for capturing this output during genesis Armada runs,
and storing it in a configurable path on the genesis node, so that it
can later be consumed by for example by the node exporter text file
collector [2]. The number of Armada attempts to collect metrics for
is also configurable.

[0]: https://airship-armada.readthedocs.io/en/latest/operations/metrics.html#exporting
[1]: https://review.opendev.org/#/c/668980/
[2]: https://github.com/prometheus/node_exporter#textfile-collector

Change-Id: Ifd5b9d351204541595b1aadf1f06b16ab54308b0