CoreDNS probe refactoring and version uplift
This is uplift for CoreDNS to version 1.6.2

Upstream CoreDNS image has no tools inside like wget/dig and can't
be used as is because pod probes will fail. Coredns pod has
Liveness/Readiness probes which are just a shell script to run
wget/dig to determine that CoreDNS is functional. So, decided
to add tools for probes in promenade image and do refactoring.

New endpoints for health check are running in side-car:
/externalhealth - to do the same check like previous shell script,
/selfcheck - to do check of the health of side-car itself.

Main container should be pointed to check endpoint provided by
side-car container.

Change-Id: Ib7fcf309b6cc34a86eeeec6e2109988cfa862955