Fix coredns checks
The readiness probe and helm test currently rely on
the return code of `dig`, which doesn't fail for
NXDOMAIN, which means they are not checking that
the names are actually resolved. This moves to using
`host` instead which does check this.

This also removes the checks for kubernetes etcd
domain names, since that doesn't get deployed until
after coredns.

Change-Id: I0b459f52663c936ed4b8b216614c5b4824a0713f