fix - pep8 and image build gates
This commit updates the Pegleg pep8 gate to use Ubuntu 18.04 (bionic) as
the base Ubuntu image for the job. Upstream CI has a community goal to
use Ubuntu 20.04 (focal) which defaults to python3.8. By moving to
python3.8, some of the pip packages need to be updated but until pip
conflicts are resolved, it is difficult to do so.

Additionally, update the deb-docker path to fix the docker image build
jobs.

Change-Id: I263c646f9fbf14cb65267df586c524de3da2878f