Update cleanup.sh developer script
Modifies the script that cleans up the local environment between runs to
allow the removal of /var/lib/kubelet/pods. The previous script hung up
on volume-subpaths (in addition to io~secret and empty-dir, which were
explicitly handled). This just unmounts everything before the files are
removed in remove_files.

Change-Id: I52cbe7d0f493336c1362ca3d3988038a1a15bcef