Stop using kube-apiserver insecure-port
The tiller container in the armada bootstrap pod relies on the insecure
port that kube-apiserver once listened on by default. The kube-apiserver
ability to serve on an insecure port, deprecated since v1.10, has been
removed in v1.20. [0]

This change updates the armada bootstrap pod to use the secure port
instead.

0: https://github.com/kubernetes/kubernetes/pull/95856

Change-Id: I6a37fa4e7f97c7aaa3cd0f61b56214483a7dc217