Add "labels" to apiserver-webhook deployment
Defines a "labels" variable and adds it to the deployment itself. In
addition, reuses said labels variable to replace other parts of the
deployment in which this was repeated. This will clean up the chart
a bit and enable Armada itself to properly wait for certain percentages
of the deployment replicas to be ready prior to proceeding. Prior to
this change, there wasn't a way to select the apiserver-webhook
deployment via labels.

Change-Id: If1ddb4feb8cde414e76a412431222f8f608f3b18