Remove log-test pod if validation succeeds
The validation function validate_kubectl_logs, which may be executed as
part of genesis or cluster join, creates a log-test pod and checks that
the 'kubectl logs' output is correct.

These completed pods don't really need to live in the cluster beyond the
initial deployment.

This change deletes the log-test pod if the validation is successful.

Change-Id: I6ae9c55f960ea70335d1fd79380c7119dc11a5e2